Fixed a bug that was caused by display cards
Because the display cards need certain information from the card variables this was being passed into them upon creation. The problem is that those variables don't actually exist unless there is an instance of the card. This means if you try and open the player deck or discard with no cards in hand it crashes the game